% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbnomics.R
\name{dbnomics}
\alias{dbnomics}
\title{DBnomics ggplot2 theme}
\usage{
dbnomics(...)
}
\arguments{
\item{...}{Arguments to be passed to \code{theme}.}
}
\description{
\code{dbnomics} is a simple ggplot2 theme for drawing nicer graphs. We do not
recommend to use it. It has been included in the package to avoid errors
from the vignette examples.
}
\examples{
\dontrun{
library(ggplot2)

rdb("IMF", "WEO", query = "France current account balance percent") \%>\%
  ggplot(aes(x = period, y = value, color = series_name)) +
  geom_line(size = 1.2) +
  geom_point(size = 2) +
  dbnomics()
}
}
