% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb_by_dimensions.R
\name{rdb_by_dimensions}
\alias{rdb_by_dimensions}
\title{Download DBnomics data using dimensions codes.}
\usage{
rdb_by_dimensions(provider_code, dataset_code, dimensions = NA,
  api_base_url = "https://api.next.nomics.world")
}
\arguments{
\item{provider_code}{Character string. DBnomics code of the provider.}

\item{dataset_code}{Character string. DBnomics code of the dataset.}

\item{dimensions}{Character string (single quote). DBnomics
code of one or several dimensions in the specified dataset.}

\item{api_base_url}{Character string. DBnomics API link.}
}
\value{
A data frame.
}
\description{
\code{rdb_by_dimensions} downloads data series from
\href{https://next.nomics.world/}{DBnomics}.
}
\details{
This function gives you access to hundreds of millions data series from
 \href{https://api.next.nomics.world/}{DBnomics API} (documentation about
  the API can be found \href{https://api.next.nomics.world/apidocs}{here}).
  The code of each series is given on the
  \href{https://next.nomics.world/}{DBnomics website}.
}
\examples{
# Fetch one value of one dimension from dataset 'Unemployment rate' (ZUTN) of AMECO provider:
df1 <- rdb_by_dimensions('AMECO','ZUTN','{"geo": ["ea12"]}')

# Fetch two values of one dimension from dataset 'Unemployment rate' (ZUTN) of AMECO provider:
df2 <- rdb_by_dimensions('AMECO','ZUTN','{"geo": ["ea12", "dnk"]}')

# Fetch several values of several dimensions from dataset 'Doing business' (DB) of World bank:
df3 <- rdb_by_dimensions('world-bank','DB','{"country": ["DZ", "BT", "PE"],"indicator": ["IC.DCP.BQCI","IC.REG.COST.PC.ZS"]}')

# Fetch one dataset 'Exports and imports by Member States of the EU/third countries'
 (namq_10_exi) of Eurostat provider:
df4 <- rdb_by_dimensions('Eurostat','namq_10_exi')
}
\seealso{
\code{\link{rdb_by_url}}, \code{\link{rdb_by_codes}}
}
