% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb_by_url.R
\name{rdb_by_url}
\alias{rdb_by_url}
\title{Download DBnomics data using url.}
\usage{
rdb_by_url(url_code)
}
\arguments{
\item{url_code}{Character string. DBnomics url of the search.}
}
\value{
A data frame.
}
\description{
\code{rdb_by_url} downloads data series from
\href{https://next.nomics.world/}{DBnomics}.
}
\details{
This function gives you access to hundreds of millions data series from
 \href{https://api.next.nomics.world/}{DBnomics API} (documentation about
  the API can be found \href{https://api.next.nomics.world/apidocs}{here}).
  The url is given on the
  \href{https://next.nomics.world/}{DBnomics website}.
}
\examples{
# Fetch one series from dataset 'Unemployment rate' (ZUTN) of AMECO provider:
df1 <- rdb_by_url('https://api.next.nomics.world/AMECO/ZUTN?series_codes=EA19.1.0.0.0.ZUTN')

# Fetch two series from dataset 'Unemployment rate' (ZUTN) of AMECO provider:
df2 <- rdb_by_url('https://api.next.nomics.world/AMECO/ZUTN?series_codes=EA19.1.0.0.0.ZUTN,DNK.1.0.0.0.ZUTN')

# Fetch one dataset 'Exports and imports by Member States of the EU/third countries'
 (namq_10_exi) of Eurostat provider:
df3 <- rdb_by_url('https://api.next.nomics.world/Eurostat/namq_10_exi')
}
\seealso{
\code{\link{rdb_by_codes}} \code{\link{rdb_by_dimensions}}
}
