% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb.R
\name{rdb}
\alias{rdb}
\title{Download DBnomics data}
\usage{
rdb(provider_code, dataset_code, series_code = NA,
  api_base_url = "https://api.next.nomics.world")
}
\arguments{
\item{provider_code}{Character string. DBnomics code of the provider.}

\item{dataset_code}{Character string. DBnomics code of the dataset.}

\item{series_code}{Character string or vector of character strings. DBnomics
code of one or several series in the specified dataset.}

\item{api_base_url}{Character string. DBnomics API link.}
}
\value{
A data frame.
}
\description{
\code{rdb} downloads data series from
\href{https://next.nomics.world/}{DBnomics}.
}
\details{
This function gives you access to hundreds of millions data series from
 \href{https://api.next.nomics.world/}{DBnomics API} (documentation about
  the API can be found \href{https://api.next.nomics.world/apidocs}{here}).
  The code of each series is given on the
  \href{https://next.nomics.world/}{DBnomics website}.
}
\examples{
# Fetch one series from dataset 'Unemployment rate' (ZUTN) of AMECO provider:
df1 <- rdb('AMECO','ZUTN','EA19.1.0.0.0.ZUTN')

# Fetch two series from dataset 'Unemployment rate' (ZUTN) of AMECO provider:
df2 <- rdb('AMECO','ZUTN',c('EA19.1.0.0.0.ZUTN','DNK.1.0.0.0.ZUTN'))

# Fetch one dataset 'Exports and imports by Member States of the EU/third countries'
 (namq_10_exi) of Eurostat provider:
df3 <- rdb('Eurostat','namq_10_exi')
}
