% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdb_last_updates.R
\name{rdb_last_updates}
\alias{rdb_last_updates}
\title{Download informations about the last DBnomics updates.}
\usage{
rdb_last_updates(all = FALSE,
  use_readLines = getOption("rdbnomics.use_readLines"),
  curl_config = getOption("rdbnomics.curl_config"))
}
\arguments{
\item{all}{Logical (default \code{FALSE}). If \code{TRUE}, then the entire
dataset of the last updates is retrieved.}

\item{use_readLines}{Logical (default \code{FALSE}). If \code{TRUE}, then
the data are requested and read with the base function \code{readLines} i.e.
through the default R internet connection. This can be used to get round the
error \code{Could not resolve host: api.db.nomics.world}.}

\item{curl_config}{Curl_handle or list (default \code{NULL}). If not
\code{NULL}, it is used for setting up a specific proxy connection. This
setup is passed to the function \code{curl_fetch_memory} of the package
\pkg{curl}. If it is a \code{curl_handle} object then it is considered to
be the argument \code{handle} of \code{curl_fetch_memory}. In the case of a
list, the names of the object elements are the names of the arguments of
\code{curl_fetch_memory}. It means that \code{curl_config = h} is
equivalent to \code{curl_config = list(handle = h)}. For
\code{curl_fetch_memory} arguments see
\code{\link[curl]{curl_fetch}}.
For available curl options see \code{\link[curl]{curl_options}},
\code{names(curl_options())} and
\href{https://curl.haxx.se/libcurl/c/curl_easy_setopt.html}{libcurl}.}
}
\value{
A \code{data.frame} or a \code{data.table}.
}
\description{
\code{rdb_last_updates} downloads informations about the last updates from
\href{https://db.nomics.world/}{DBnomics}.
}
\details{
By default, the function returns a \code{data.frame} (or a \code{data.table})
containing the last 100 updates from
\href{https://db.nomics.world/}{DBnomics} with additional informations.
}
\examples{
\dontrun{
rdb_last_updates()

rdb_last_updates(all = TRUE)
}
}
\seealso{
\code{\link{rdb_providers}}
}
